Rename experimental-products -> products
