+++ plugins/woocommerce-admin/client/tasks/fills/products/footer.tsx
