+++ plugins/woocommerce-admin/client/tasks/fills/products/icon/lightbulb_24px.js
