+++ plugins/woocommerce-admin/client/tasks/fills/products/icon/link_24px.js
