+++ plugins/woocommerce-admin/client/tasks/fills/products/icon/widgets_24px.js
