+++ plugins/woocommerce-admin/client/tasks/fills/products/index.scss
