+++ plugins/woocommerce-admin/client/tasks/fills/products/stack.tsx
