+++ plugins/woocommerce-admin/client/tasks/fills/products/test/footer.tsx
